requirements.txt
typeguard
flask
python-dotenv
requests
openai
#mindee